some cleanup on the hub
