+++ common/models/config.go
@@ -30,3 +30,15 @@ type ProviderData struct {
 	ProviderData ProviderDB `json:"provider"`
 	DeviceData   []*Device  `json:"device_data"`
 }
+
+type HubConfig struct {
+	HostAddress          string `json:"host_address"`
+	Port                 string `json:"port"`
+	MongoDB              string `json:"mongo_db"`
+	SeleniumGridInstance string `json:"selenium_grid_instance"`
+	AdminUsername        string `json:"admin_username"`
+	AdminPassword        string `json:"admin_password"`
+	AdminEmail           string `json:"admin_email"`
+	OSTempDir            string `json:"-"`
+	UIFilesTempDir       string `json:"-"`
+}
